refactor: step1결과 사용해서 파일 내용 가져오는 로직 분리
