+++ apps/server/src/projects/projects.service.ts
@@ -9,6 +9,7 @@ import {
 import { GitignoreMatcherService } from './services/gitignore-matcher.service';
 import { ProjectRepository } from './repository/project.repository';
 import { NcloudStorageService } from '../storage/ncloud-storage.service';
+import { Step1Result } from '../ai/types/ai.types';
 
 const unlink = promisify(fs.unlink);
 
@@ -89,5 +90,50 @@ export class ProjectsService {
         folders: structureData.folders || [],
       },
     };
+  }
+
+  /**
+   * Step1 결과에서 요청한 주요 파일들을 NCloud에서 다운로드해 추출합니다.
+   * @param projectId 프로젝트 ID
+   * @param step1Result Step1 분석 결과
+   * @returns 파일 경로 → 파일 내용 매핑
+   */
+  async extractMainFiles(
+    projectId: string,
+    step1Result: Step1Result,
+  ): Promise<Record<string, string>> {
+    if (
+      !step1Result.project_main_files?.length ||
+      step1Result.project_main_files.length === 0
+    ) {
+      return {};
+    }
+
+    const paths = step1Result.project_main_files.map((item) => item.file_path);
+    const objectKey = NcloudStorageService.objectKeyForProject(projectId);
+    let tmpPath: string | null = null;
+
+    try {
+      tmpPath = await this.ncloudStorage.downloadToTempFile(objectKey);
+      const fileContents = await this.zipParser.extractPaths(tmpPath, paths);
+      return Object.fromEntries(fileContents);
+    } finally {
+      if (tmpPath) {
+        try {
+          await unlink(tmpPath);
+        } catch (error) {
+          console.error(`임시 파일 삭제 실패: ${tmpPath}`, error);
+        }
+      }
+    }
+  }
+
+  /**
+   * 프로젝트의 NCloud ZIP 파일을 삭제합니다.
+   * @param projectId 프로젝트 ID
+   */
+  async deleteProjectZip(projectId: string): Promise<void> {
+    const objectKey = NcloudStorageService.objectKeyForProject(projectId);
+    await this.ncloudStorage.deleteObject(objectKey);
   }
 }
